Corrects ExternalName-typed services reporting missing Endpoints
This change moves the decision to error on missing endpoints/pods for a
service from k8s/client.go to linter/svc.go. In the cases where the svc
has no ClusterIP (where it is explicitly None, or not allowed such as
in the ExternalName svc type) we'll skip the check for Endpoints.